Support parent folder

+++ download_and_clean.py
@@ -1,7 +1,5 @@
 import argparse
-import json
 import multiprocessing
-import re
 from concurrent.futures import ProcessPoolExecutor
 from itertools import repeat
 
@@ -9,23 +7,24 @@ import putiopy
 import os
 from tqdm import tqdm
 
-from libgen_to_txt.files import get_file_path, download_folder, download_folder_locally, delete_file_locally
+from libgen_to_txt.files import get_file_path, download_folder, download_folder_locally, delete_file_locally, \
+    get_parent_id
 from libgen_to_txt.marker.convert import process_folder_marker
 from libgen_to_txt.metadata import batch_write_metadata
 from libgen_to_txt.naive.convert import process_batch_files_naive
 from libgen_to_txt.settings import settings
 
 
-def process_single_libgen_chunk(torrent_info, conversion_lock, no_download, max_workers=settings.CONVERSION_WORKERS):
+def process_single_libgen_chunk(torrent_info, conversion_lock, no_download, no_delete, max_workers=settings.CONVERSION_WORKERS):
     num, url = torrent_info
 
     client = putiopy.Client(settings.PUTIO_TOKEN, timeout=15, use_retry=True)
-    sel_file = get_file_path(num, client)
+    parent_folder_id = get_parent_id(client)
+    sel_file = get_file_path(num, client, parent_folder_id)
 
     if not sel_file:
-        sel_file = download_folder(url, num, client, no_download)
+        sel_file = download_folder(url, num, client, parent_folder_id, no_download)
         if not sel_file:
-            print(f"Failed to download {num}, or took too long")
             return
 
     stored_path = download_folder_locally(sel_file.name)
@@ -54,13 +53,18 @@ def process_single_libgen_chunk(torrent_info, conversion_lock, no_download, max_
         f.write(sel_file.name)
 
     # Delete files from remote and local
-    sel_file.delete()
-    delete_file_locally(sel_file.name)
+    if not no_download:
+        sel_file.delete()
+
+    if not no_delete:
+        delete_file_locally(sel_file.name)
 
 
-def try_process_single_libgen_chunk(torrent_info, lock, no_download=False):
+def try_process_single_libgen_chunk(torrent_info, lock, options):
     try:
-        process_single_libgen_chunk(torrent_info, lock, no_download)
+        no_download = options["no_download"]
+        no_delete = options["no_delete"]
+        process_single_libgen_chunk(torrent_info, lock, no_download, no_delete)
     except Exception as e:
         print(f"Failed to process {torrent_info}: {e}")
 
@@ -69,7 +73,8 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Download and process libgen")
     parser.add_argument("--max", type=int, default=None, help="Maximum number of chunks to process, for testing")
     parser.add_argument("--workers", type=int, default=settings.DOWNLOAD_WORKERS, help="Number of workers to use when downloading")
-    parser.add_argument("--no_download", type=int, action="store_true", help="Only process what already exists on the seedbox", default=False)
+    parser.add_argument("--no_download", action="store_true", help="Only process what already exists on the seedbox", default=False)
+    parser.add_argument("--no_delete", action="store_true", help="Do not delete files locally", default=False)
     args = parser.parse_args()
 
     os.makedirs(settings.BASE_STORAGE_FOLDER, exist_ok=True)
@@ -92,7 +97,12 @@ if __name__ == "__main__":
 
     m = multiprocessing.Manager()
     lock = m.Lock()
-    with ProcessPoolExecutor(max_workers=args.workers) as pool:
-        tqdm(pool.map(try_process_single_libgen_chunk, torrent_urls, repeat(lock), repeat(args.no_download), chunksize=1), total=len(torrent_urls))
-    pool.shutdown()
+    args_dict = vars(args)
+    if args.workers == 1:
+        for torrent_info in tqdm(torrent_urls):
+            try_process_single_libgen_chunk(torrent_info, lock, args_dict)
+    else:
+        with ProcessPoolExecutor(max_workers=args.workers) as pool:
+            tqdm(pool.map(try_process_single_libgen_chunk, torrent_urls, repeat(lock), repeat(args_dict), chunksize=1), total=len(torrent_urls))
+        pool.shutdown()
